feat: show daily items in item/note-row
Shows individual daily items in indexes, so that viewing an index
exposes some of the daily content. Includes support for showing only
daily children that match a set of tags.

This has bugged me for a while I'm happy to expose content that is
somewhat buried in the daily notes.