fix: break state on spaces
Gave up trying to use pretty print here.

Unfortunately, this is going to explode the line numbers in the repo -
maybe /public should be a submodule, or kicked into some other repo?

The real problem here is trying to review updates to posts, so get some
idea of what changes were made - ideally we could use git for this
directly.